# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "NetflixAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Subscription": {
      "get": {
        "tags": [
          "Subscription"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subscription"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subscription"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subscription"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subscription"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subscription"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Subscription"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Subscription"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Subscription"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Subscription"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Subscription"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Subscription"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Subscription"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          }
        }
      }
    },
    "/api/Subscription/{id}": {
      "get": {
        "tags": [
          "Subscription"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Subscription"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Subscription"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Subscription"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Subscription"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Subscription"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Subscription"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Subscription"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Subscription"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/SubtitleContent": {
      "get": {
        "tags": [
          "SubtitleContent"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubtitleContent"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubtitleContent"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubtitleContent"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubtitleContent"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubtitleContent"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SubtitleContent"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleContent"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleContent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleContent"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleContent"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleContent"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleContent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SubtitleContent"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubtitleContent"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubtitleContent"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SubtitleContent"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SubtitleContent"
                }
              }
            }
          }
        }
      }
    },
    "/api/SubtitleContent/{id}": {
      "get": {
        "tags": [
          "SubtitleContent"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SubtitleContent"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubtitleContent"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubtitleContent"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SubtitleContent"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SubtitleContent"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "SubtitleContent"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleContent"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleContent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleContent"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleContent"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleContent"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleContent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "SubtitleContent"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Subscription": {
        "type": "object",
        "properties": {
          "subscriptionId": {
            "type": "integer",
            "format": "int64"
          },
          "subscriptionName": {
            "type": "string",
            "nullable": true
          },
          "subscriptionCost": {
            "type": "number",
            "format": "float"
          },
          "qualitiesAllowed": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SubtitleContent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "episodeId": {
            "type": "integer",
            "format": "int64"
          },
          "languageId": {
            "type": "integer",
            "format": "int64"
          },
          "subtitleLocation": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}